app/requirements.txt
aiohappyeyeballs==2.4.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
Deprecated==1.2.14
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.114.2
frozenlist==1.4.1
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
limits==3.13.0
loguru==0.7.2
multidict==6.1.0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pydantic==2.9.1
pydantic_core==2.23.3
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
tomli==2.0.1
typing_extensions==4.12.2
uvicorn==0.30.6
wrapt==1.16.0
yarl==1.11.1
